SNMDEMO: New decompilation
Since NUTCracker doesn't work with the floppy disk version of the interactive Sam and Max demo (due to the resource file using the .SM* extension), I had to use the CD version instead. Some of the scripts don't seem to have been decompiled properly. I assume it's because this is v7, which hasn't been fully tested.